chore : 불필요한 주석 제거 - 코드 내 불필요한 설명 및 중복 로직 제거로 가독성 향상
