+++ apps/commerce-api/src/main/java/com/loopers/domain/brand/BrandDomainCreateRequest.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.Size;
 
 /**
  * 브랜드 도메인 생성 요청 DTO
- * <p>
+ * 
  * 도메인 레이어에서 사용하는 브랜드 생성 요청 정보입니다.
  * Application Layer에서 Domain Layer로 전달되는 데이터 구조입니다.
  *
