+++ apps/commerce-api/src/main/java/com/loopers/domain/brand/BrandService.java
@@ -17,7 +17,7 @@ import jakarta.validation.Valid;
 
 /**
  * 브랜드 도메인 서비스
- * <p>
+ * 
  * 브랜드 도메인의 비즈니스 로직을 처리합니다.
  * 단일 책임 원칙에 따라 브랜드 Repository에만 의존합니다.
  *
@@ -43,7 +43,7 @@ public class BrandService {
     /**
      * 모든 활성 브랜드를 조회합니다.
      *
-     * <p>캐시 배치 갱신 등에서 사용됩니다.</p>
+     * 캐시 배치 갱신 등에서 사용됩니다.
      *
      * @return 모든 활성 브랜드 목록
      */
@@ -98,7 +98,7 @@ public class BrandService {
 
     /**
      * 브랜드를 등록합니다.
-     * <p>
+     * 
      * 브랜드 등록은 단일 도메인 작업이므로 도메인 서비스에서 트랜잭션 처리합니다.
      *
      * @param request 브랜드 생성 요청 정보
