+++ apps/commerce-api/src/main/java/com/loopers/domain/like/LikeService.java
@@ -12,7 +12,7 @@ import lombok.RequiredArgsConstructor;
 
 /**
  * 좋아요 도메인 서비스
- * <p>
+ * 
  * 좋아요 도메인의 비즈니스 로직을 처리합니다.
  * 좋아요와 상품 간의 협력을 통해 좋아요 카운트를 관리합니다.
  *
@@ -38,11 +38,11 @@ public class LikeService {
 
     /**
      * 좋아요를 등록하거나 복원합니다 (Upsert).
-     * <p>
+     * 
      * - 좋아요 관계가 없으면: 새로 생성하고 카운트 증가
      * - 삭제된 좋아요가 있으면: 복원하고 카운트 증가
      * - 활성 좋아요가 있으면: 기존 엔티티 반환 (카운트 변경 없음 - 중복 방지)
-     * <p>
+     * 
      * 좋아요 카운트는 DB 원자적 연산(UPDATE 쿼리)으로 처리하여 동시성을 보장합니다.
      *
      * @param user    사용자 엔티티
@@ -71,7 +71,7 @@ public class LikeService {
 
     /**
      * 좋아요를 취소합니다 (소프트 삭제).
-     * <p>
+     * 
      * 좋아요를 삭제하고 상품의 좋아요 카운트를 감소시킵니다.
      * 좋아요 카운트는 DB 원자적 연산(UPDATE 쿼리)으로 처리하여 동시성을 보장합니다.
      *
