+++ apps/commerce-api/src/main/java/com/loopers/domain/product/ProductMVService.java
@@ -22,8 +22,8 @@ import lombok.extern.slf4j.Slf4j;
 /**
  * 상품 Materialized View 서비스
  *
- * <p>MV 테이블 조회 및 배치 동기화를 담당하는 서비스입니다.
- * 상품, 브랜드, 좋아요 정보를 통합하여 조회 성능을 최적화합니다.</p>
+ * MV 테이블 조회 및 배치 동기화를 담당하는 서비스입니다.
+ * 상품, 브랜드, 좋아요 정보를 통합하여 조회 성능을 최적화합니다.
  *
  * @author hyunjikoh
  * @since 2025. 11. 27.
@@ -98,12 +98,12 @@ public class ProductMVService {
      * MV 테이블을 원본 테이블과 동기화합니다.
      * 배치 스케줄러에서 주기적으로 호출됩니다.
      *
-     * <p>✅ 개선된 로직:</p>
-     * <ul>
-     *   <li>단일 쿼리로 변경된 데이터만 조회 (Product + Brand + Like 조인)</li>
-     *   <li>불필요한 개별 쿼리 제거</li>
-     *   <li>마지막 배치 시간 이후 변경된 데이터만 처리</li>
-     * </ul>
+     * ✅ 개선된 로직:
+     *
+     *   단일 쿼리로 변경된 데이터만 조회 (Product + Brand + Like 조인)
+     *   불필요한 개별 쿼리 제거
+     *   마지막 배치 시간 이후 변경된 데이터만 처리
+     *
      *
      * @return 배치 업데이트 결과
      */
