+++ apps/commerce-api/src/main/java/com/loopers/domain/product/ProductService.java
@@ -14,7 +14,7 @@ import jakarta.validation.Valid;
 
 /**
  * 상품 도메인 서비스
- * <p>
+ * 
  * 상품 도메인의 비즈니스 로직을 처리합니다.
  * 단일 책임 원칙에 따라 상품 Repository에만 의존합니다.
  *
@@ -87,7 +87,7 @@ public class ProductService {
 
     /**
      * 상품을 등록합니다.
-     * <p>
+     * 
      * 상품 등록은 단일 도메인 작업이므로 도메인 서비스에서 트랜잭션 처리합니다.
      *
      * @param request 상품 생성 요청 정보
@@ -103,7 +103,7 @@ public class ProductService {
 
     /**
      * 상품 엔티티의 재고를 차감합니다.
-     * <p>
+     * 
      * 이미 조회된 상품 엔티티의 재고를 차감할 때 사용합니다.
      *
      * @param product  재고를 차감할 상품 엔티티
@@ -118,7 +118,7 @@ public class ProductService {
 
     /**
      * 상품 재고를 원복합니다.
-     * <p>
+     * 
      * 주문 취소 시 차감된 재고를 다시 복구합니다.
      *
      * @param productId 상품 ID
