+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/cache/CacheKeyGenerator.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 
 /**
  * 캐시 키 생성기
- * <p>
+ * 
  * Hot/Warm/Cold 전략별 캐시 키 생성
  * - product:detail:{productId}
  * - product:ids:{strategy}:{brandId}:{page}:{size}:{sort}
