+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/product/ProductMVBatchScheduler.java
@@ -24,11 +24,11 @@ import lombok.extern.slf4j.Slf4j;
 /**
  * ProductMaterializedView 배치 스케줄러
  *
- * <p>MV 테이블 동기화 및 Hot 캐시 갱신을 통합 관리합니다.</p>
- * <ul>
- *   <li>2분마다: MV 테이블 동기화</li>
- *   <li>50분마다: Hot 캐시 갱신 (인기순 상품, 브랜드별 인기순)</li>
- * </ul>
+ * MV 테이블 동기화 및 Hot 캐시 갱신을 통합 관리합니다.
+ *
+ *   2분마다: MV 테이블 동기화
+ *   50분마다: Hot 캐시 갱신 (인기순 상품, 브랜드별 인기순)
+ *
  */
 @Component
 @RequiredArgsConstructor
@@ -79,8 +79,8 @@ public class ProductMVBatchScheduler {
     /**
      * Hot 캐시 갱신 배치 작업 (50분마다)
      *
-     * <p>배치 갱신으로 캐시 스탬피드 방지</p>
-     * <p>ProductMVRepository를 직접 사용하여 likeCount 정렬 보장</p>
+     * 배치 갱신으로 캐시 스탬피드 방지
+     * ProductMVRepository를 직접 사용하여 likeCount 정렬 보장
      */
     @Scheduled(fixedRate = 50 * 60 * 1000, initialDelay = 60 * 1000)
     public void refreshHotCache() {
