+++ apps/commerce-api/src/test/java/com/loopers/domain/like/LikeIntegrationTest.java
@@ -40,7 +40,6 @@ import lombok.extern.slf4j.Slf4j;
 
 /**
  * LikeFacade 통합 테스트
- * <p>
  * 좋아요 등록 및 취소 기능을 파사드 레벨에서 검증합니다.
  * 각 도메인 서비스를 통해 트랜잭션이 적용된 데이터 저장을 수행합니다.
  *
